+++ ClimateExplorer.Core/DataPreparation/Model/ChartableDataPoint.cs
@@ -3,6 +3,6 @@
 public class ChartableDataPoint
 {
     public string? Label { get; set; }
-    public float? Value { get; set; }
+    public double? Value { get; set; }
     public string? BinId { get; set; }
 }
